requirements.txt
pandas>=1.5.0
numpy>=1.21.0
requests>=2.28.0
openpyxl>=3.0.0
xlrd>=2.0.0
scikit-learn>=1.1.0
torch>=1.12.0
transformers>=4.20.0
rouge-score>=0.1.0
nltk>=3.7
spacy>=3.4.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0
openai>=0.27.0
anthropic>=0.5.0
ollama>=0.1.0
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
